Add files via upload
This project is done as 1st year student in university. In this project , i created a website to track the health and growth of infants that can acess by parents, Mid wives,doctors and admin. Parent can get the BMI value of there kid by inputting height and weight of their kid, can get the vaccine dates, menus to feed their kids, and doctors can evaluvate the baby in good health and growth.