Add files via upload
This is a project done by Visual Studio. In this project we develop a system to report a threat happen to woman for higher authorities like Sri Lanka Police, Women and child care Authority and for a Physician or councilor if there a mental abuse happen to women. The main objective is to enhance to security for women